Switch gossip mode when dc and rack are specified
When CASSANDRA_DC or CASSANDRA_RACK are specified, the gossip mode
must also be changed to GossipingPropertyFileSnitch. It is a common
error for people to miss this step and be confused by the results.
This patch automatically makes this change for the user if the
environment variable CASSANDRA_ENDPOINT_SNITCH is not set (i.e. if
it is left at it's default value of SimpleSnitch).

Fixes #57